requirements.txt
gunicorn
Flask==2.0.1
Flask-Cors==3.0.10
google-cloud-aiplatform==1.8.0
google-auth==2.3.0
Pillow==8.3.2
numpy==1.19.5
